# ServicioCalculadorGradoDeConfianza/Domain/Confianza/GradoDeConfianza.cs
namespace ServicioCalculadorGradoDeConfianza.Domain.Confianza
{
    public enum GradoDeConfianza
    {
        NoConfiable,
        ConReservas,
        ConfiableNivel1,
        ConfiableNivel2
    }
}