# ServicioCalculadorGradoDeConfianza/SwaggerDocumentation.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "ServicioCalculadorGradoDeConfianza",
      "version": "1.0"
    },
    "paths": {
      "/gradoDeConfianza/usuario": {
        "post": {
          "tags": [
            "GradoDeConfianza"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/GradoDeConfianzaResponseUsuario"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GradoDeConfianzaResponseUsuario"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GradoDeConfianzaResponseUsuario"
                  }
                }
              }
            }
          }
        }
      },
      "/gradoDeConfianza/comunidad": {
        "post": {
          "tags": [
            "GradoDeConfianza"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComunidadRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComunidadRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ComunidadRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/GradoDeConfianzaResponseComunidad"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GradoDeConfianzaResponseComunidad"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GradoDeConfianzaResponseComunidad"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Comunidad": {
          "type": "object",
          "properties": {
            "gradoDeConfianza": {
              "type": "string",
              "nullable": true
            },
            "miembros": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Usuario"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "ComunidadRequest": {
          "type": "object",
          "properties": {
            "comunidad": {
              "$ref": "#/components/schemas/Comunidad"
            },
            "incidentes": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Incidente"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "GradoDeConfianza": {
          "enum": [
            0,
            1,
            2,
            3
          ],
          "type": "integer",
          "format": "int32"
        },
        "GradoDeConfianzaResponseComunidad": {
          "type": "object",
          "properties": {
            "comunidad": {
              "$ref": "#/components/schemas/Comunidad"
            },
            "nuevoPuntaje": {
              "type": "number",
              "format": "float"
            },
            "gradoDeConfianzaActual": {
              "$ref": "#/components/schemas/GradoDeConfianza"
            }
          },
          "additionalProperties": false
        },
        "GradoDeConfianzaResponseUsuario": {
          "type": "object",
          "properties": {
            "usuario": {
              "$ref": "#/components/schemas/Usuario"
            },
            "nuevoPuntaje": {
              "type": "number",
              "format": "float"
            },
            "gradoDeConfianzaActual": {
              "$ref": "#/components/schemas/GradoDeConfianza"
            }
          },
          "additionalProperties": false
        },
        "Incidente": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "fechaApertura": {
              "type": "string",
              "format": "date-time"
            },
            "usuarioReportador": {
              "$ref": "#/components/schemas/Usuario"
            },
            "fechaCierre": {
              "type": "string",
              "format": "date-time"
            },
            "usuarioAnalizador": {
              "$ref": "#/components/schemas/Usuario"
            },
            "servicioAfectado": {
              "$ref": "#/components/schemas/PrestacionDeServicio"
            }
          },
          "additionalProperties": false
        },
        "PrestacionDeServicio": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            }
          },
          "additionalProperties": false
        },
        "Usuario": {
          "type": "object",
          "properties": {
            "nombre": {
              "type": "string",
              "nullable": true
            },
            "apellido": {
              "type": "string",
              "nullable": true
            },
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "puntosDeConfianza": {
              "type": "number",
              "format": "float"
            }
          },
          "additionalProperties": false
        },
        "UsuarioRequest": {
          "type": "object",
          "properties": {
            "usuario": {
              "$ref": "#/components/schemas/Usuario"
            },
            "incidentes": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Incidente"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      }
    }
  }